pyproject.toml
[project]
name = "rag-assgn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.3.1",
    "bs4>=0.0.2",
    "colorlog>=6.10.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "numpy>=2.4.4",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.56.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.44.0",
]
